Configuracion de rutas, junto con mejora visual con Antd y configuraciones de Firebase

+++ Vite+Pinia+FireBase/src/stores/database.js
@@ -3,6 +3,7 @@ import {defineStore} from 'pinia'
 import {db} from '../firebaseConfig'
 import {auth} from '../firebaseConfig'
 import { nanoid } from 'nanoid'
+import router from '../router'
 
 export const useDatabaseStore = defineStore('database', {
     state: () => ({
@@ -11,20 +12,25 @@ export const useDatabaseStore = defineStore('database', {
     }),
     actions:{
         async getUrls(){
-            if(this.documents.length !== 0){
+            console.log('getUrls ejecutado')
+            if (!auth.currentUser){
+                console.log('No hay usuarios')
                 return
-            }
+            } 
             this.loadingDoc = true
+            this.documents = []
             try {
-                const q = query(collection(db,'urls'), where('user','==',auth.currentUser.uid))
-                const queryShapshot = await getDocs(q)
-                queryShapshot.forEach(doc => {
-                    console.log(doc.id, doc.data())
+                const q = query(collection(db,'urls'),where('doc.user','==',auth.currentUser.uid))
+                const querySnapshot = await getDocs(q)
+                console.log("UID usado:", auth.currentUser.uid);
+                console.log("respuesta", querySnapshot);
+                querySnapshot.forEach(doc => {
                     this.documents.push({
                         id: doc.id,
                         ...doc.data()
                     })
                 })
+                console.log(this.documents)
             } catch (error) {
                 console.error(error)
             } finally {
@@ -32,23 +38,31 @@ export const useDatabaseStore = defineStore('database', {
             }
         },
         async addUrl(name) {
+            this.loadingDoc = true
             try {
-                const doc = { name:name, short:nanoid(6), user:auth.currentUser.uid}
+                const new_doc = { name, short:nanoid(6), user:auth.currentUser.uid}
                 const docRef = await addDoc(collection(db, 'urls'), {
-
+                    doc:new_doc
+                })
+                this.documents.push({
+                    id:docRef.id,
+                    doc:doc
                 })
             } catch (error) {
                 console.error(error)
+            } finally {
+                this.loadingDoc = false
             }
         },
         async deleteUrl(id) {
+            if (!auth.currentUser) return 
             this.loadingDoc = true
             try {
                 await deleteDoc(doc(db, 'urls', id))
                 this.documents = this.documents.filter(item => item.id !== id)
-                console.log(dele)
+                console.log(`Documento eliminado correctamente: ${id}`)
             } catch (error) {
-                console.log(error.message)
+                console.error('Error al eliminar:', error.message)
             } finally {
                 this.loadingDoc = false
             }
@@ -57,7 +71,7 @@ export const useDatabaseStore = defineStore('database', {
             this.loadingDoc = true
             try {
                 const docSnap = await getDoc(doc(db, 'urls', id))
-                return docSnap.data().name
+                return docSnap.data().doc.name
             } catch (error) {
                 console.log(error.message)
             } finally {
@@ -67,10 +81,10 @@ export const useDatabaseStore = defineStore('database', {
         async updateUrl(id, name) {
             this.loadingDoc = true
             try {
-                await updateDoc(doc(db, 'urls', id), { name })
+                await updateDoc(doc(db, 'urls', id), { 'doc.name':name })
                 this.documents = this.documents.map(item =>
-                item.id === id ? { ...item, name } : item
-                )
+                item.id === id ? { ...item, doc:{ ...item.doc,name}} : item)
+                router.push('/')
             } catch (error) {
                 console.log(error.message)
             } finally {
